Create a pom.xml file to handle ant dependencies
To build TNT4J using Ant requires a bunch of dependencies that need to be stored in the lib/ directory.  A simple pom.xml with the maven dependency plugin will allow these dependencies to be easily managed.  

To use run mvn install first followed by the ant command as usual.  Versions of the required packages can be updated in the pom as usual